\name{Rshadow-package}
\alias{Rshadow-package}
\alias{Rshadow}
\docType{package}
\title{
  High-performance optimization for R
}
\description{
  SHADOw stands for Sparse Hessian Automatic Differentiation for Optimization FloWs. Rshadow is
  the R front end for the C++ SHADOw library. Rshadow implements second-order optimization
  of arbitrarily complex objective functions.
}
\details{
  ...
}
\author{
Jean-Philippe Rasigade

Maintainer: Jean-Philippe Rasigade, <jean-philippe.rasigade@univ-lyon1.fr>
}
\references{
  Current SHADOw repository:
  https://github.com/rasigadelab/shadow

  ADOL-C version with the LIVARH algorithm (not used in SHADOw code but useful as reference):
  https://github.com/CSCsw/LivarH

  Wang, M., Gebremedhin, A. & Pothen, A. Capitalizing on live variables: new algorithms for efficient Hessian computation via automatic differentiation. Math. Prog. Comp. 8, 393–433 (2016). https://doi.org/10.1007/s12532-016-0100-3
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
